Fix v16 example, restructure headers (#98)
* Overview how to integrate libocpp outside of EVerest
* Fix share path in 1.6 charge point example
* Remove unnecessary certificates from readme
* - updated method documentation of charge_point.hpp
- removed some unused fields
* Add a error message when the chargepoint example can't find the config
Fixed search path for logging.ini
* Update README concerning running of the OCPP 1.6 example
* Add utils, encapsulating case insensitive string comparison
* Enable option in nlohmann json for forward declaration support
* Add missing utils.cpp
* Move Message and Call/CallResult/CallError related types to their own file
* Restructure DateTime operators
* Restructure Powermeter to_ and from_json conversions
* Some header cleanup
* Use nlohmann json forward declaration in messages
* Silence unused parameter warning in messages that have no parameters
* Replace boost::filesystem with std::filesystem
* Accept ChargePoint config as string instead of nlohmann json
* Cleanup header
* Remove usage of removed register_set_max_current_callback
* cleanup charge_point header by introducing charge_point_impl
* Fixed some compiler warnings + formatting
* Fix database handler includes
* Fix update_clock_aligned_meter_values_interval
* Fix typo
* Fix duplicate definitions
* Fixed std::optional usage
* Fixed some compiler warnings
* Re-generated OCPP 1.6 messages with code generator
* Re-generated OCPP 2.0.1 messages with code generator
* Fix optional array of enum in code generator
* Also use std::optional in OCPP 1.6 example
* Cleanup includes
* Update README to reflect recent changes to charger point API
* fixed some typos and added some more information to the readme

---------

+++ lib/ocpp/v16/charge_point_state_machine.cpp
@@ -141,21 +141,21 @@ void ChargePointStates::reset(std::map<int, v16::AvailabilityType> connector_ava
 
 void ChargePointStates::submit_event(int connector_id, FSMEvent event) {
     const std::lock_guard<std::mutex> lck(state_machines_mutex);
-    if (connector_id > 0 && connector_id < this->state_machines.size()) {
+    if (connector_id > 0 && (size_t)connector_id < this->state_machines.size()) {
         this->state_machines.at(connector_id).handle_event(event);
     }
 }
 
 void ChargePointStates::submit_error(int connector_id, const ChargePointErrorCode& error_code) {
     const std::lock_guard<std::mutex> lck(state_machines_mutex);
-    if (connector_id > 0 && connector_id < state_machines.size()) {
+    if (connector_id > 0 && (size_t)connector_id < state_machines.size()) {
         state_machines.at(connector_id).handle_fault(error_code);
     }
 }
 
 ChargePointStatus ChargePointStates::get_state(int connector_id) {
     const std::lock_guard<std::mutex> lck(state_machines_mutex);
-    if (connector_id > 0 && connector_id < this->state_machines.size()) {
+    if (connector_id > 0 && (size_t)connector_id < this->state_machines.size()) {
         return state_machines.at(connector_id).get_state();
     } else if (connector_id == 0) {
         return ChargePointStatus::Available;
